feat: Add subfield pruning and complex type function decomposition (#5)
Summary:
Performs a pass on the input PlanNode tree to determine which subfields of complex type columns and intermediate results are accessed. This fills in maps from PlanNode to column, subfield set pairs for control columns and payload columns. Control columns affect selection of rows, payload columns are along for the ride and are candidates for late materialization.

Control columns are filter inputs, join keys and grouping keys.

We annotate PlanNodes and TypedExprs with the subfield information. the TypedExpr case is relevant for functions that return complex types.

We implement a optimizer::FunctionRegistry for subfield and second order function metadata. When a function that returns complex types is used inside a subfield path, the subfield paths can serve to determine which subfields of the function's input are accessed. This is the case for functions like transform and transform_values.

Additionally, genie functions (DPP functions that operate on all/most columns of the row and return another isomorphic row can be made decomposable. Specifying an explode function does this as follows: The function gets the complete set of subfields that are accessed in the return value. The explode function then looks at the starts of the paths and for each distinct start it understands, it produces an expression. This expression is then substituted for the path where the path is referenced.

For example, select genie(id_list_features).id_list_features[11] from features applies the path id_list_features[11] to the genie. The genie exploder sees this and can return an expression. Also, we automatically know that only [11] is accessed in id_list_features in features. This last fact is inferred based on a mapping from produced subfields to corresponding input arguments of the genie. See FunctionMetadata.

A side effect of this analysis is that we automatically prune unreferenced columns from projections and aggregates. Previously this was not explicitly done.

To implement subfield support, we add a Path object. This is a deduplicated RAII set of distinct paths, each with its id. as with Exprs, comparison is by pointer identity. We add a Path::skyline function which removes all paths that are occluded by a shorter path sharing a common prefix with the longer one.

Additionally, this diff refactors BitSet as separate from PlanObjectSet and fixes lifetimes of velox::variant.

Pull Request resolved: https://github.com/facebookexperimental/verax/pull/5

Reviewed By: Yuhta

Differential Revision: D70673650

Pulled By: oerling

fbshipit-source-id: ab55c04046cf6f16e727b1714ed43ab7670b4d41