Support map as struct and pushdown of subfields (#6)
Summary:
Allows specifying a set of columns that are expected to be flat maps and will be accessed as structs s long as the whole map is not accessed.

Extracts column skylines and groups all projections to extract single subfields inside or near the leaf table scan if subfield projection is enabled.

Adds a an extra argument for output type to sampling so that map as struct can be supported.  The dataColumns must be the type in the file and the output types can specify a struct instead of a map.

Pull Request resolved: https://github.com/facebookexperimental/verax/pull/6

Reviewed By: Yuhta

Differential Revision: D71742794

Pulled By: oerling

fbshipit-source-id: eb3f92d65ade32e8291241349a46f201271d34f1