feat: Parallelize wide expressions (#7)
Summary:
Adds an options for parallelizing data independent projections.

Adds a ParallelProject Operator using operator extensibility .

Verax can have a target parallelism for projections. If so, we divide a set of projection expressions into levels using a depth first traversal. If we encounter a multiply referenced Expr, we place it below the lowest referencing level.

After this, we detect one or more multiply referenced (common subexpression) borders. the border is the set of leafmost multiply referenced Exprs such that these do not contain unplaced multiply referenced exprs.

the first border is placed in a projection node. The constituent Exprs are marked placed. We repeat until there are no more multiply referenced Exprs.

At this point we detect a parallelizable border. This border consists of Exprs whose cost including children is over a target threshold. After this, we have top level Exprs, like constructors left but the bulk of the workk is below the parallel border.

We place the expressions in te parallel border. Then we make a finall single threaded projection to stitch the results together. This consists of the Exprs above the parallel border.

All but the last projection are parallel and guarantee non-overlapping Expr trees.

A parallel projection has groups of Exprs that are evaluated together. Each such group is independently schedulable. Additionally it has identity projected columns that are not subject to lazy loading.

When we place the parallel border, we also load all lazies so that lazy loading does columns independently. These are passed through the common subexpression levels because these are typically less wide. We place the bulk f lazy loads and projections into one large reorderable project to expose maximum parallelism.

The ParallelProject operator is like FilterProject without the filter and just runs independent groups of expressions using AsyncSource.

The test adds a facility for making sythetic pseudorandom projection expressions. This can be used for testing and benchmarking parallel projection.

Pull Request resolved: https://github.com/facebookexperimental/verax/pull/7

Reviewed By: Yuhta

Differential Revision: D72118453

Pulled By: oerling

fbshipit-source-id: 7d6ec7e12f3f3c7b18c2f0afad84539cfb7fa07f