fix: Fix join graph diamonds (#12)
Summary:
In Q9 and Q5 there are diamonds where a table is added to a plan through one join edge.  There are more join edges between the added table and the plan made so far. Therefore, make a new composite edge combining all the inner joins between the new table and the already placed tables.

There are cases (Q11) where expr deduplication is not correct because n_name = xx occurs twice and refers to different nation tables in each case. The dedup map must be from func + ExprCP... to ExprCP and not from TypedExprPtr to ExprCP. Like this the args of the function call to dedup are correctly scoped.

Also, dedupping literals so that the same literal has the same ExprCP so that x + 1 and x + 1 dedup. They would not if the literal where not physically the same.

Pull Request resolved: https://github.com/facebookexperimental/verax/pull/12

Reviewed By: Yuhta

Differential Revision: D75033743

Pulled By: oerling

fbshipit-source-id: 8bfb6b3bab950751ffbfcf8f8e8750ea518353a3