fix: unbreak logical plan optimization for multi-connector queries (#132)
Summary:
Pull Request resolved: https://github.com/facebookexperimental/verax/pull/132

For queries which reference tables residing on multiple connectors, we need a way to specify to the optimizer which connector to query for table metadata

Previously, with the physical input interfaces, we passed a fully-qualified table name (`[catalog].[schema].[query]`). Now there is a dedicated field in logical scan nodes for the connector ID

Modifying the schema lookup in Verax to respect this separately specified catalog by passing it explicitly. If the plan creator created the scan with a fully-qualified table name as before, the catalog specified there is compared against the separate connector ID parameter

Reviewed By: mbasmanova

Differential Revision: D79820585

fbshipit-source-id: 8ed2ec4f02d0bf06f4564e0e36eceba48ddeef27