verax(feat): move findTable to smart pointer return type (#227)
Summary:
Pull Request resolved: https://github.com/facebookexperimental/verax/pull/227

was hard to manage the lifecycle of tables when the Optimizer (or other findTable callers) could dereference the pointer at any time. e.g. if a second concurrent query determines the table information is stale, or planning is stuck and the table is pushed out of cache, the server can segfault

Changing the return type to shared_ptr<const Table> (ConnectorTablePtr) and updating all the callers

Also, adding a set in the Optimizer to capture a reference to all the Tables used in Optimization, to prevent reclaimation prior to Optimization completion

Reviewed By: oerling

Differential Revision: D80525688

fbshipit-source-id: 1e83c9ab97a2d828fdff42f5d712d240f98cbd88