feat(connectors): Switch dataPath -> tablePath API (#422)
Summary:
Pull Request resolved: https://github.com/facebookexperimental/verax/pull/422

The actual construction of the table path should be implementation-specific. e.g., it should be possible for a variety of Hive datastore to switch the storage path depending on the table. For example, if a Hive implementation would like to obfuscate the table path during write

Since all current callers of this API only use the path to construct a table directory, it's easy to pull this logic into the API itself

Reviewed By: pansatadru

Differential Revision: D83078815

fbshipit-source-id: 92a95eda24bc6d83717929edf59ab2157bc316e8