fix: Aggregations with expressions shared between grouping keys and aggregates (#435)
Summary:
Fix optimization for aggregations that share expressions between grouping keys and aggregates. For example,

> SELECT a + b, count(a + b) FROM t GROUP BY 1

The error was due to adding shared expressions to `renames_`  before translating all expressions for the aggregation node.

This fix is to record renames in a separate map while processing aggregation node, then move it to `renames_`.

Pull Request resolved: https://github.com/facebookexperimental/verax/pull/435

Reviewed By: Yuhta

Differential Revision: D83368338

Pulled By: mbasmanova

fbshipit-source-id: cc079d6fdfb60075fc1aa9c1f08aad5b7f886bb9