feat: Add support for TableWrite (#479)
Summary:
Extend the Optimizer to support logical plans with TableWriteNode as a root. Allows to create plans that append data to existing tables.

The input to TableWriteNode may produce only a subset of the columns in a table. The rest of the columns will be populated with default values. The connector is responsible for specifying default values. By default, these are assumed to be NULL.

ToGraph creates a separate DT for TableWrite. This allows to simply the rest of the code as DT with TableWrite doesn't contain anything else.

The optimizer doesn't compute the cost of TableWrite. Since every candidate plan must include TableWrite in the end, the cost of write doesn't affect cost comparison for the candidates.

Extend `PlanAndStats` to add `finishWrite` callback that invokes `ConnectorMetadata::finishWrite` or `abortWrite`.

Change `ConnectorMetadata::finishWrite` to return number of rows.
And actually compute number of rows for LocalHiveConnectorMetadata::finishWrite.

Change `ConnectorMetadata::abortWrite` to make it noexcept.
And actually make `LocalHiveConnectorMetadata::abortWrite` noexcept.

Extend LocalRunner to run finishWrite callback and construct vector with number of written rows.

Added `shared_from_this` to `axiom::connector::Table` because it's tricky to pass `shared_ptr<Table>` from `ToGraph` to `ToVelox`. TODO: Consider other way of doing this without `shared_from_this`, but also without additional hashmap/vector/etc.

Follow-ups:
1. Add tests for default column values with non-null default value.
2. Implement WriteKind::kCreate, kUpdate, kDelete.
4. Add support for writing partitioned tables.
5. Update LocalHiveConnectorMetadata to use Velox FileSystem instead of Linux-specific APIs.

Pull Request resolved: https://github.com/facebookincubator/axiom/pull/479

Reviewed By: hdikeman

Differential Revision: D84291111

Pulled By: mbasmanova

fbshipit-source-id: 5c92c1b3fab044ab404eaa3d88c1c7eeafcaf9c0