feat: Add optimizer support for writing partitioned tables (#511)
Summary:

In multi-node setup, when writing to partitioned table (bucketed Hive tables), check if input is already partitioned in a compatible way. If not, add a shuffle.

In single-node multi-threaded setup, add local exchange before partitioned write.

Check the table layout to determine whether write is partitioned or not. Tables with multiple layouts are not supported yet. To support these, we'll need to introduce layout-for-write or primary layout or similar.

Reviewed By: xiaoxmeng

Differential Revision: D84718419