refactor: Introduce PlanCost (#550)
Summary:
Pull Request resolved: https://github.com/facebookincubator/axiom/pull/550

Cost struct is used to represent a cost of a single RelationOp as well as a cost of a Plan (a tree of RelationOp). This is confusing. Introduce PlanCost to represent the cost of a Plan and limit usage of Cost to representing the cost of a RelationOp.

Add optimizer option to disable exploring plans with reducing semi joins. Use this option in TpchPlanTest.q09 as a workaround for https://github.com/facebookincubator/axiom/issues/530

Reviewed By: xiaoxmeng

Differential Revision: D85342911

fbshipit-source-id: d80f3ed84d0330edf92cd066e2d12b6ffda49daf