fix: Import of reducing semi joins in q9
Summary:
Partial fix for https://github.com/facebookincubator/axiom/issues/530

Adapted from Orri's https://github.com/facebookincubator/axiom/pull/551

When we import an already placed table or tables into a build side for cardinality reduction as a semijoin:

The colums on the right side of the semijoin must not be added to downstreamColumns. A semijoin (exists) or antijoin (not exists) does not produce columns, except maybe a mark. Also, if the join has filters, the columns in the filter that come from the right side are not downstream columns.

Co-authored-by: oerling

Differential Revision: D85464350