feat: Add support for non-correlated subqueries used in filters (#568)
Summary:
Pull Request resolved: https://github.com/facebookincubator/axiom/pull/568

Add minimal support for non-correlated subqueries used in a filter. Enables TPC-H q15, 16 and 18.

Add logic to process non-correlated subqueries found in a filter.

IN <subquery> expression is handled by creating a new DT for the subquery and adding a semi-join edge to the current DT. The IN predicate is then replaced with a 'mark' column produced by the join.

Other <subquery> expressions are handled by creating a new DT for the subquery and replacing the expression with the only column produced by the DT.

Reviewed By: xiaoxmeng

Differential Revision: D85625835

fbshipit-source-id: ecdea52e3a72828b470295e421455c33ccdb04bb