refactor: Remove connector::TableKind (#589)
Summary:
This type varies across connectors and is not used by core components like LogicalPlan, Optimizer, or Runner. Since it's an unnecessary abstraction with a flawed design, we should remove it now before it becomes more entrenched.

For an example in SereneDB we have
- Temporary tables
- RocksDB tables and indexes
- Search Indexes
- System catalog tables (pg_type, pg_class, etc)

They're all have different `axiom::connector::Table` implementation, because `numRows` required to be implemented in a different way

Pull Request resolved: https://github.com/facebookincubator/axiom/pull/589

Reviewed By: jagill

Differential Revision: D86110072

Pulled By: kevinwilfong

fbshipit-source-id: 1a1d896163865b078aad19f727258c7c30354b1c