fix: Join fanout in Optimization::joinByHashRight (#632)
Summary:
Pull Request resolved: https://github.com/facebookincubator/axiom/pull/632

"a left b" is equivalent to "b left a". Hence, we expect that cardinality of both joins would be the same. However, due to a bug in 'joinByHashRight' the cardinality of right join was 10x more than left join.

Here is how it went for q13 (roughly):

- |a| is 100
- |b| is 1000
- a-to-b fanout (lrFanout) is 10 (each row in 'a' has 10 matches in 'b')
- b-to-a fanout (rlFanout) is 1 (each row in 'b' has a match in 'a')

Join relation representing 'a LEFT b' was created with fanout 10 and reported cardinality of |a| * 10 = 1000.

Join relation representing 'b RIGHT a' was created with fanout 10 and reported cardinality of |b| * 10 = 10000.

The inflated cardinality caused inflated cost of the aggregation and made RIGHT join plan appear more costly than LEFT join plan.

bypass-github-export-checks

Reviewed By: xiaoxmeng

Differential Revision: D87081215

fbshipit-source-id: df50415b1dfcdd944992ccda8d3ee074e8537959