feat: Prune join build output
Summary:
In selecting the columns to produce from a bushy build side, consider the build tables placed, so that we only produce columns needed downstream of the join, not columns that are internal to the bushy build.

In considering a join candidate, when the candidate has multiple tables connected by inner join, consider all edges from the placed plan to any of the other tables, add the equalities in these edges to the join edge in the candidate. This means that the derived table that is made from the inner join in the candidate will be joined to the already placed plan on all the relevant columns. Also call addExtraEdges after creating bushy builds with inner joins and existences.

In forming the composite edge, combine the fanouts as they would be combined if the joins were consecutive. The combined fanout is the lower of the min of the fanouts and the product of the fanouts.

Change the arguments of JoinEdge::setFanouts to be (left-to-right, right-to-left), as this is the conventional order and is also used in this way in DerivedTable.cpp.

Add plan verification for TPC-H q5.

Differential Revision: D88400329