feat: Multi-statement Parsing (#701)
Summary:
Pull Request resolved: https://github.com/facebookincubator/axiom/pull/701

Extend the PrestoParser to have a method to split text into multiple statements. This will be used to support multi-statement execution in the axiom CLI. This will provide compatibility with the presto CLI which can execute multiple delimited statements. We will generally match the presto CLI behavior as a target.

Reviewed By: pansatadru

Differential Revision: D88758186

fbshipit-source-id: 59d003ff2f422935846b376d3210b71f5c7dafcd