misc: Respect Linux cgroups restrictions when sizing thread pools (#712)
Summary:
Pull Request resolved: https://github.com/facebookincubator/axiom/pull/712

This change replaces uses of std::thread::hardware_concurrency with
folly::hardware_concurrency the latter of which always respects
container restrictions while the former does not.  This ensures that
when running inside of a Linux container with restrictions on the
available processors, thread pools and other data structures will not
be over-provisioned.

Reviewed By: hdikeman, mbasmanova

Differential Revision: D87913058

fbshipit-source-id: 66c2012a20c5845c116ec890bc221836fe65b5ec