build: Fix build errors on Mac (#710)
Summary:
Replace nnL with nnLL or (int64_t)nn.

The difference between the L and LL suffixes is that L indicates a long int, which is not guaranteed to be 64-bit, while LL indicates a long long int, which is guaranteed to be at least 64-bit.

This fixes compilation errors. There are still linking errors.

Pull Request resolved: https://github.com/facebookincubator/axiom/pull/710

Reviewed By: pratikpugalia

Differential Revision: D89564994

Pulled By: kKPulla

fbshipit-source-id: 85f48dcede55578b150c260f19b1af4a036a0134