feat: Introduce hidden columns (#725)
Summary:
Pull Request resolved: https://github.com/facebookincubator/axiom/pull/725

Some connectors support hidden columns that do not appear in the output of SELECT * FROM t queries, but can be queried directly.

For example, Hive connector supports $path and $bucket hidden columns that contain the file path and bucket number of the row. The values of these columns are included in the splits. The connector populates these using constant vectors on the fly.

https://prestodb.io/docs/current/connector/hive.html#extra-hidden-columns

This change adds 'hidden' flag to axiom::connector::Column class to allow marking the column as hidden.

Reviewed By: amitkdutta

Differential Revision: D89712385

fbshipit-source-id: 21b6d93b39df5bfb1a4d84f7a78f9ec7e9d3b831